requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
requests
python-dotenv